requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
python-docx==1.1.2
openai==1.57.2
email-validator==2.2.0